+++ src/assets/css/Cards.css
@@ -16,8 +16,6 @@
 }
 
 .card-icon {
-  width: 3rem;
-  height: 3rem;
   color: #e38071;
 }
