+++ src/components/Homepage/cards/Cards.jsx
@@ -1,9 +1,6 @@
 import React from "react";
-import CardTotalRider from "./CardTotalRider.jsx";
-import AVGCommuters from "./AVG-Commuters.jsx";
-import AVGrevenue from "./AVG-Revenue.jsx";
+import CardItem from "./CardItem.jsx";
 import "../../../assets/css/Card.css";
-import AVGVerifiedRider from "./AVG-VerifiedRider.jsx";
 import { Row, Col } from "reactstrap";
 
 const Cards = () => {
@@ -11,28 +8,28 @@ const Cards = () => {
     <>
       <Row>
         <Col className="mb-lg-2 mb-md-3" xl={3} lg={3} md={6} sm={10}>
-          <CardTotalRider
+          <CardItem
             cardIcon={"bicycle"}
             text={"Total Riders"}
             numericText={"2.5k"}
           />
         </Col>
         <Col className="mb-lg-2 mb-md-3" xl={3} lg={3} md={6} sm={10}>
-          <CardTotalRider
+          <CardItem
             cardIcon={"user"}
             text={"Monthly Commuters"}
             numericText={"12k"}
           />
         </Col>
         <Col className="mb-lg-2 mb-md-3" xl={3} lg={3} md={6} sm={10}>
-          <CardTotalRider
+          <CardItem
             cardIcon={"money"}
             text={"Monthly Revenue"}
             numericText={"125k"}
           />
         </Col>
         <Col className="mb-lg-2 mb-md-3" xl={3} lg={3} md={6} sm={10}>
-          <CardTotalRider
+          <CardItem
             cardIcon={"certificate"}
             text={"Monthly verified"}
             numericText={"1.2k"}
